Add example of changing binary device class
The "Binary Sensor" page describes all of the different device classes for a binary sensor.  If you Google search how to change a device class, all of the forum posts and videos refer back to when everything was in configuration.yaml.  I'd like to add this small section giving an example of how to set the device class through the GUI.